+++ CMLibrary/src/main/java/com/boha/coursemaker/util/ImageTask.java
@@ -48,7 +48,7 @@ public class ImageTask {
 		}
 		@Override
 		protected void onPostExecute(Integer res) {
-			if (res.intValue() > 0) {
+			if (res > 0) {
 				bitmapListener.onError();
 				return;
 			}
